Bao branch

+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/OrderController.cs
@@ -47,7 +47,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetOrderById")]
+        [HttpGet("GetOrderById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -95,15 +95,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await OrderExists(id)) { return NotFound("Order not found!"); }
+
             string command = $"DELETE FROM Orders WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> OrderExists(int id)
         {
             string command = $"SELECT * FROM Orders WHERE Id = {id}";
