+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/OrderItemController.cs
@@ -44,7 +44,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetOrderItemById")]
+        [HttpGet("GetOrderItemById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -89,15 +89,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await OrderItemExists(id)) { return NotFound("Order Item not found!"); }
+
             string command = $"DELETE FROM OrderItem WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> OrderItemExists(int id)
         {
             string command = $"SELECT * FROM OrderItem WHERE Id = {id}";
