+++ Backend/TestForASPWebAPI/TestForASPWebAPI/Controllers/ProductImageController.cs
@@ -43,7 +43,7 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // GET api/<ValuesController>/5
-        [HttpGet("GetProductImageById")]
+        [HttpGet("GetProductImageById/{id}")]
         public async Task<IActionResult> Get(int id)
         {
             DBController dbController = DBController.GetInstance();
@@ -87,15 +87,18 @@ namespace TestForASPWebAPI.Controllers
         }
 
         // DELETE api/<ValuesController>/5
-        [HttpDelete("Delete")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        public async Task<IActionResult> Delete(int id)
         {
+            if (!await ProductImageExists(id)) { return NotFound("Product Image not found!"); }
+
             string command = $"DELETE FROM ProductImage WHERE Id = {id}";
             DBController dbController = DBController.GetInstance();
             dbController.DeleteData(command);
-            return;
+
+            return NoContent();
         }
-        [HttpGet("Exists")]
+        [HttpGet("Exists/{id}")]
         public async Task<bool> ProductImageExists(int id)
         {
             string command = $"SELECT * FROM ProductImage WHERE Id = {id}";
